[refactor]114685 MHV-Medications - Seperate helper functions into seperate files
