+++ src/applications/mhv-medications/util/helpers/extractContainedResource.js
@@ -1,0 +1,19 @@
+import { isArrayAndHasItems } from './isArrayAndHasItems';
+
+/**
+ * Extract a contained resource from a FHIR resource's "contained" array.
+ * @param {Object} resource a FHIR resource (e.g. AllergyIntolerance)
+ * @param {String} referenceId an internal ID referencing a contained resource
+ * @returns the specified contained FHIR resource, or null if not found
+ */
+export const extractContainedResource = (resource, referenceId) => {
+  if (resource && isArrayAndHasItems(resource.contained) && referenceId) {
+    // Strip the leading "#" from the reference.
+    const strippedRefId = referenceId.substring(1);
+    const containedResource = resource.contained.find(
+      containedItem => containedItem.id === strippedRefId,
+    );
+    return containedResource || null;
+  }
+  return null;
+};
