+++ src/applications/mhv-medications/util/helpers/generateTextFile.js
@@ -1,0 +1,15 @@
+/**
+ * Download a text file
+ * @param {String} content text file content
+ * @param {String} fileName name for the text file
+ */
+export const generateTextFile = (content, fileName) => {
+  const blob = new Blob([content], { type: 'text/plain' });
+  const url = window.URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = fileName;
+  a.click();
+  window.URL.revokeObjectURL(url);
+  a.remove();
+};
