+++ src/applications/mhv-medications/util/helpers/getReactions.js
@@ -1,0 +1,14 @@
+/**
+ * @param {Object} record
+ * @returns {Array of Strings} array of reactions
+ */
+export const getReactions = record => {
+  const reactions = [];
+  if (!record || !record.reaction) return reactions;
+  record.reaction.forEach(reaction => {
+    reaction.manifestation.forEach(manifestation => {
+      reactions.push(manifestation.text);
+    });
+  });
+  return reactions;
+};
