+++ src/applications/mhv-medications/util/helpers/isArrayAndHasItems.js
@@ -1,0 +1,7 @@
+/**
+ * @param {Any} obj
+ * @returns {Boolean} true if obj is an array and has at least one item
+ */
+export const isArrayAndHasItems = obj => {
+  return Array.isArray(obj) && obj.length;
+};
